ion: Fixup custom compat ioctl
If the task that is calling the compat ioctl is not a compat
task, pass through the arguments to the regular ioctl. The ioctl
sizes may be different because of type size differences so the
compat ioctl cmd may not match the incoming cmd. Also, the types
of offset and length are ints not longs so change the types accordingly.

Change-Id: I32585e07a556f18dda81c64683d061b48c773311
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>